refactor: ToyCrudService::deleteBy에 예외처리 추가

+++ app/src/main/java/com/codesoom/assignment/application/ToyCrudService.java
@@ -51,6 +51,10 @@ public class ToyCrudService implements ToyCreateService, ToyShowService, ToyUpda
 
     @Override
     public void deleteBy(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ProductNotFoundException(id);
+        }
+
         repository.deleteById(id);
     }
 }
